Add log in and out functionality with navbar
The functionality was added that the navbar shows the log in dropdown if the user isn't logged in and it shows the profile dropdown if the user is logged in. The profile dropdown now includes the ability for a user to log out and when they do it redirects them to the home page.
The side drawer that appears when the navbar is too small also has this functionality updated.
I changed some styling on the register and login pages to fix some spacing and width issues.
The only thing that is not done is getting the navbar to show correctly right after they log in without the user having to refresh. I am currently working on that.