Small fixes
Fixed buttons on home screen so when the screen is small the text doesn't overlap out of the buttons.
Fixed footer to make text smaller to fit properly when screen is smaller.
Made about us page work so text fit on image and when screen is too small it populates below image instead.
Fixed contact us form so that when the screen gets smaller it takes up more of the screen so the button doesn't have the text outside of it.